requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
phonenumbers>=8.12.23,<8.13
django-phonenumber-field>=5.2.0,<5.3
pytest>=6.2.5,<7.0
pytest-django>=4.4.0,<5.0
djangorestframework-simplejwt>=4.7.2,<4.8
drf-spectacular>=0.15.1,<0.16
reportlab
django-filter>=21.1,<22.0